+++ prtm/models/__init__.py
@@ -1,3 +1,5 @@
+import importlib
+
 from prtm.models.antiberty import *
 from prtm.models.dmpfold2 import *
 from prtm.models.eigenfold import *
@@ -7,6 +9,7 @@ from prtm.models.genie import *
 from prtm.models.igfold import *
 from prtm.models.omegafold import *
 from prtm.models.openfold import *
+from prtm.models.pifold import *
 from prtm.models.protein_generator import *
 from prtm.models.proteinmpnn import *
 from prtm.models.proteinsolver import *
@@ -14,7 +17,6 @@ from prtm.models.rfdiffusion import *
 from prtm.models.rosettafold import *
 from prtm.models.se3_diffusion import *
 
-import importlib
-if importlib.util.find_spec('pyrosetta'):
+if importlib.util.find_spec("pyrosetta"):
     from prtm.models.protein_seq_des import *
 del importlib
