+++ prtm/models/pifold/__init__.py
@@ -1,0 +1,2 @@
+import prtm.models.pifold.config as pifold_config
+from prtm.models.pifold.modeling import PiFoldForInverseFolding
