--- prtm/models/pifold/methods_utils.py
@@ -1,47 +1,0 @@
-from collections.abc import Mapping, Sequence
-
-import numpy as np
-import torch
-
-
-def cuda(obj, *args, **kwargs):
-    """
-    Transfer any nested conatiner of tensors to CUDA.
-    """
-    if hasattr(obj, "cuda"):
-        return obj.cuda(*args, **kwargs)
-    elif isinstance(obj, Mapping):
-        return type(obj)({k: cuda(v, *args, **kwargs) for k, v in obj.items()})
-    elif isinstance(obj, Sequence):
-        return type(obj)(cuda(x, *args, **kwargs) for x in obj)
-    elif isinstance(obj, np.ndarray):
-        return torch.tensor(obj, *args, **kwargs)
-
-    raise TypeError("Can't transfer object type `%s`" % type(obj))
-
-
-def loss_smoothed(S, log_probs, mask, num_classes=20, weight=0.1):
-    """Negative log probabilities"""
-    S_onehot = torch.nn.functional.one_hot(
-        S, num_classes=num_classes
-    ).float()  # [4, 463] --> [4, 463, 20]
-
-    # Label smoothing
-    S_onehot = S_onehot + weight / float(S_onehot.size(-1))
-    S_onehot = S_onehot / S_onehot.sum(
-        -1, keepdim=True
-    )  # [4, 463, 20]/[4, 463, 1] --> [4, 463, 20]
-
-    loss = -(S_onehot * log_probs).sum(-1)
-    loss_av = torch.sum(loss * mask) / torch.sum(mask)
-    return loss, loss_av
-
-
-def loss_nll(S, log_probs, mask):
-    """Negative log probabilities"""
-    criterion = torch.nn.NLLLoss(reduction="none")
-    loss = criterion(
-        log_probs.contiguous().view(-1, log_probs.size(-1)), S.contiguous().view(-1)
-    ).view(S.size())
-    loss_av = torch.sum(loss * mask) / torch.sum(mask)
-    return loss, loss_av
